fix: restore vial keymap
