--- keyboards/durgod/k3x0/keymaps/vial/readme.md
@@ -1,16 +1,0 @@
-# The default keymap for Durgod Taurus K3X0 (K310, K320, etc.) without backlight
-
-Supports both ANSI and ISO layouts.
-
-Layer 0 : Standard layout (Windows)
-
-Layer 1 : Media control and Windows lock key
-- Reusing Durgod's Original Media Control for Fn + F1 ~ Fn + F7
-- Fn + Windows to toggle Windows lock key functionality
-
-## Windows key lock
-
-You can hold down Fn + Windows key to disable the Windows key while in locked mode.
-The 'Lock' LED indicates if the Windows key is locked.
-
-This is similar to the stock K310/K320 Windows key lock functionality.
